requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==23.0.0
requests==2.32.3
